fix: properly read stderr progress
