+++ .changes/command-stdio-return.md
@@ -1,0 +1,6 @@
+---
+"tauri": patch
+"api": patch
+---
+
+The process Command API stdio lines now includes the trailing `\r`.
