+++ .changes/io-read-line-util.md
@@ -1,0 +1,5 @@
+---
+"tauri-utils": patch
+---
+
+Added the `io` module with the `read_line` method.
