+++ core/tauri-utils/Cargo.toml
@@ -32,6 +32,7 @@ json5 = { version = "0.4", optional = true }
 json-patch = "0.2"
 glob = { version = "0.3.0", optional = true }
 walkdir = { version = "2", optional = true }
+memchr = "2.4"
 
 [target."cfg(target_os = \"linux\")".dependencies]
 heck = "0.4"
